Move repositories to database package
