+++ src/main/java/com/arte/backend/model/database/repository/UserRepository.java
@@ -1,4 +1,4 @@
-package com.arte.backend.repository;
+package com.arte.backend.model.database.repository;
 
 import com.arte.backend.model.database.entity.UserData;
 import org.springframework.data.jpa.repository.JpaRepository;
