+++ src/main/java/com/arte/backend/service/statistics/UserStatisticsService.java
@@ -3,8 +3,8 @@ package com.arte.backend.service.statistics;
 import com.arte.backend.model.database.entity.UserData;
 import com.arte.backend.model.database.entity.UserStatistics;
 import com.arte.backend.model.statistics.UserStatisticsDataModel;
-import com.arte.backend.repository.UserRepository;
-import com.arte.backend.repository.UserStatisticsRepository;
+import com.arte.backend.model.database.repository.UserRepository;
+import com.arte.backend.model.database.repository.UserStatisticsRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.stereotype.Service;
